Move integration framework into packages
