+++ tests/integration/framework/kill/kill_posix.go
@@ -14,20 +14,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package kill
 
 import (
 	"os"
+	"os/exec"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 )
 
-func (c *Command) interrupt(t *testing.T) {
+func interrupt(t *testing.T, cmd *exec.Cmd) {
 	// TODO: daprd does not currently gracefully exit on a single interrupt
 	// signal. Remove once fixed.
-	assert.NoError(t, c.cmd.Process.Signal(os.Interrupt))
+	assert.NoError(t, cmd.Process.Signal(os.Interrupt))
 	time.Sleep(time.Millisecond * 300)
-	assert.NoError(t, c.cmd.Process.Signal(os.Interrupt))
+	assert.NoError(t, cmd.Process.Signal(os.Interrupt))
 }
